requirements.txt
openai>=1.93.0
pyaudio>=0.2.13
wave>=0.0.2
python-dotenv>=1.0.0
pyttsx3>=2.90
requests>=2.32.4
